Add debugging event and update collectible logic
Introduced Debugging_Event script for testing and debugging purposes. Updated CollectibleManager to invoke a new event when a specific number of collectibles is reached, and added related UnityEvent fields. Modified the In Game scene to include the new Collectible Manager GameObject and adjusted some object properties. Also added a new scene to the build settings.